Update draft-turner-ccmib.md
Based on feedback received from A. Farrel (11/19/2019), a re-examination of the asymmetric key table (and related tables) was done.  A few errors were discovered, notably the existence of a single read-create column in the cAsymKeyTable (the RowStatus column is read-write), and the existence of a RowStatus column in the cCertSubAltNameTable (all other columns are read-only).